Migrate test to respective test file
